Add Prometheus metrics support
Addresses #15.

Contributes towards epimorphics/hmlr-linked-data#57

This commit adds support for collecting metrics relating to calls to the
data API. We add an dependency for an `instrumenter`, which defaults to
Rails' `ActiveSupport::Notifications`. By publishing events via the
notifier, apps which include `ds-api-ruby` can be configured to collect
those events and feed them to a Prometheus store.

While setting up the dependency injection for the instrumenter, I have
also added a similar hook to inject the logger. This will allow us to
improve the ability of the logger to emit correctly formulated JSON
messages.

This change also adds significantly more unit test coverage for the
elements of `service.rb`.